requirements.txt
alembic==1.4.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Login==0.5.0
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
marshmallow-sqlalchemy==0.23.1
